// Frontend/app/api/generated/model/dashboardResumenDTO.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface DashboardResumenDTO {
  areas?: number;
  carreras?: number;
  materias?: number;
  usuarios?: number;
  docentes?: number;
  administrativos?: number;
  programasTotales?: number;
  programasVigentes?: number;
  pendienteAdministracion?: number;
  pendienteDocente?: number;
  rechazadoDocente?: number;
  pendienteComisiones?: number;
  pendienteSecretaria?: number;
}